Moving database/handler and repository to external
