+++ main.go
@@ -5,10 +5,10 @@ import (
 	"log/slog"
 	"os"
 
+	"github.com/ViniAlvesMartins/tech-challenge-fiap/external/database/postgres"
+	"github.com/ViniAlvesMartins/tech-challenge-fiap/external/handler/http_server"
+	"github.com/ViniAlvesMartins/tech-challenge-fiap/external/repository"
 	"github.com/ViniAlvesMartins/tech-challenge-fiap/infra"
-	"github.com/ViniAlvesMartins/tech-challenge-fiap/src/adapter/database/postgres"
-	"github.com/ViniAlvesMartins/tech-challenge-fiap/src/adapter/inbound/handler/http_server"
-	"github.com/ViniAlvesMartins/tech-challenge-fiap/src/adapter/outbound/repository"
 	"github.com/ViniAlvesMartins/tech-challenge-fiap/src/core/service"
 	"gorm.io/gorm"
 )
